表記ミス / Modalの名称変更
